mds: big refactor of MDS class (stage 1)
Separate all the things which are logically associated with
an actual MDS rank into MDSRank.  Subsystems etc will now
have a reference to an MDSRank rather than the MDS instance
itself, limiting the scope of what they have access to.

At this stage, MDS simply inherits from MDSRank to get
all the fields, and subsystems are seeing an MDSRank*
pointer that actually points to an MDS instance.  Next
step is to make MDS encapsulate MDSRank instead
of descending from it.
