+++ src/mds/MDS.cc
@@ -60,11 +60,6 @@
 
 #include "messages/MGenericMessage.h"
 
-#include "messages/MClientRequest.h"
-#include "messages/MClientRequestForward.h"
-
-#include "messages/MMDSTableRequest.h"
-
 #include "messages/MMonCommand.h"
 #include "messages/MCommand.h"
 #include "messages/MCommandReply.h"
@@ -82,14 +77,37 @@
 #undef dout_prefix
 #define dout_prefix *_dout << "mds." << whoami << '.' << incarnation << ' '
 
+/**
+ * FIXME: These guys are weird, they're touching the global mds_lock
+ * via the MDSRank's reference to it.  For contexts which really truly
+ * related to MDSDaemon rather than MDSRank, create a different class
+ * for them to use that doesn't relate to MDSRank.
+ */
+class MDSDaemonIOContext :public MDSIOContext {
+  protected:
+    MDS *mds_daemon;
+  public:
+    MDSDaemonIOContext(MDS *m) : MDSIOContext(m), mds_daemon(m) {}
+};
+
+/**
+ * FIXME as for IOContext
+ */
+class MDSDaemonInternalContext :public MDSInternalContext {
+  protected:
+    MDS *mds_daemon;
+  public:
+    MDSDaemonInternalContext(MDS *m) : MDSInternalContext(m), mds_daemon(m) {}
+};
+
 
 // cons/des
 MDS::MDS(const std::string &n, Messenger *m, MonClient *mc) : 
+  MDSRank(mds_lock, clog, timer, mdsmap, &finisher, this, m, mc),
   Dispatcher(m->cct),
   mds_lock("MDS::mds_lock"),
   stopping(false),
   timer(m->cct, mds_lock),
-  hb(NULL),
   beacon(m->cct, mc, n),
   authorize_handler_cluster_registry(new AuthAuthorizeHandlerRegistry(m->cct,
 								      m->cct->_conf->auth_supported.empty() ?
@@ -100,24 +118,15 @@ MDS::MDS(const std::string &n, Messenger *m, MonClient *mc) :
 								      m->cct->_conf->auth_service_required :
 								      m->cct->_conf->auth_supported)),
   name(n),
-  whoami(MDS_RANK_NONE), incarnation(0),
   standby_for_rank(MDSMap::MDS_NO_STANDBY_PREF),
   standby_type(MDSMap::STATE_NULL),
   standby_replaying(false),
   messenger(m),
   monc(mc),
   log_client(m->cct, messenger, &mc->monmap, LogClient::NO_FLAGS),
-  op_tracker(cct, m->cct->_conf->mds_enable_op_tracker, 
-                     m->cct->_conf->osd_num_op_tracker_shard),
   finisher(cct),
-  osd_epoch_barrier(0),
-  sessionmap(this),
-  progress_thread(this),
   asok_hook(NULL)
 {
-
-  hb = cct->get_heartbeat_map()->add_worker("MDS");
-
   orig_argc = 0;
   orig_argv = NULL;
 
@@ -134,7 +143,7 @@ MDS::MDS(const std::string &n, Messenger *m, MonClient *mc) :
 
   mdcache = new MDCache(this);
   mdlog = new MDLog(this);
-  balancer = new MDBalancer(this);
+  balancer = new MDBalancer(this, messenger);
 
   inotable = new InoTable(this);
   snapserver = new SnapServer(this);
@@ -150,8 +159,6 @@ MDS::MDS(const std::string &n, Messenger *m, MonClient *mc) :
   
   // tick
   tick_event = 0;
-
-  req_rate = 0;
 
   last_state = want_state = state = MDSMap::STATE_BOOT;
 
@@ -195,10 +202,6 @@ MDS::~MDS() {
   
   if (messenger)
     delete messenger;
-
-  if (hb) {
-    cct->get_heartbeat_map()->remove_worker(hb);
-  }
 }
 
 class MDSSocketHook : public AdminSocketHook {
@@ -985,152 +988,6 @@ void MDS::create_logger()
   server->create_logger();
   mdcache->register_perfcounters();
 }
-
-
-
-MDSTableClient *MDS::get_table_client(int t)
-{
-  switch (t) {
-  case TABLE_ANCHOR: return NULL;
-  case TABLE_SNAP: return snapclient;
-  default: assert(0);
-  }
-}
-
-MDSTableServer *MDS::get_table_server(int t)
-{
-  switch (t) {
-  case TABLE_ANCHOR: return NULL;
-  case TABLE_SNAP: return snapserver;
-  default: assert(0);
-  }
-}
-
-
-
-
-
-
-
-
-void MDS::send_message(Message *m, Connection *c)
-{ 
-  assert(c);
-  c->send_message(m);
-}
-
-
-void MDS::send_message_mds(Message *m, mds_rank_t mds)
-{
-  if (!mdsmap->is_up(mds)) {
-    dout(10) << "send_message_mds mds." << mds << " not up, dropping " << *m << dendl;
-    m->put();
-    return;
-  }
-
-  // send mdsmap first?
-  if (mds != whoami && peer_mdsmap_epoch[mds] < mdsmap->get_epoch()) {
-    messenger->send_message(new MMDSMap(monc->get_fsid(), mdsmap), 
-			    mdsmap->get_inst(mds));
-    peer_mdsmap_epoch[mds] = mdsmap->get_epoch();
-  }
-
-  // send message
-  messenger->send_message(m, mdsmap->get_inst(mds));
-}
-
-void MDS::forward_message_mds(Message *m, mds_rank_t mds)
-{
-  assert(mds != whoami);
-
-  // client request?
-  if (m->get_type() == CEPH_MSG_CLIENT_REQUEST &&
-      (static_cast<MClientRequest*>(m))->get_source().is_client()) {
-    MClientRequest *creq = static_cast<MClientRequest*>(m);
-    creq->inc_num_fwd();    // inc forward counter
-
-    /*
-     * don't actually forward if non-idempotent!
-     * client has to do it.  although the MDS will ignore duplicate requests,
-     * the affected metadata may migrate, in which case the new authority
-     * won't have the metareq_id in the completed request map.
-     */
-    // NEW: always make the client resend!  
-    bool client_must_resend = true;  //!creq->can_forward();
-
-    // tell the client where it should go
-    messenger->send_message(new MClientRequestForward(creq->get_tid(), mds, creq->get_num_fwd(),
-						      client_must_resend),
-			    creq->get_source_inst());
-    
-    if (client_must_resend) {
-      m->put();
-      return; 
-    }
-  }
-
-  // these are the only types of messages we should be 'forwarding'; they
-  // explicitly encode their source mds, which gets clobbered when we resend
-  // them here.
-  assert(m->get_type() == MSG_MDS_DIRUPDATE ||
-	 m->get_type() == MSG_MDS_EXPORTDIRDISCOVER);
-
-  // send mdsmap first?
-  if (peer_mdsmap_epoch[mds] < mdsmap->get_epoch()) {
-    messenger->send_message(new MMDSMap(monc->get_fsid(), mdsmap), 
-			    mdsmap->get_inst(mds));
-    peer_mdsmap_epoch[mds] = mdsmap->get_epoch();
-  }
-
-  messenger->send_message(m, mdsmap->get_inst(mds));
-}
-
-
-
-void MDS::send_message_client_counted(Message *m, client_t client)
-{
-  Session *session =  sessionmap.get_session(entity_name_t::CLIENT(client.v));
-  if (session) {
-    send_message_client_counted(m, session);
-  } else {
-    dout(10) << "send_message_client_counted no session for client." << client << " " << *m << dendl;
-  }
-}
-
-void MDS::send_message_client_counted(Message *m, Connection *connection)
-{
-  Session *session = static_cast<Session *>(connection->get_priv());
-  if (session) {
-    session->put();  // do not carry ref
-    send_message_client_counted(m, session);
-  } else {
-    dout(10) << "send_message_client_counted has no session for " << m->get_source_inst() << dendl;
-    // another Connection took over the Session
-  }
-}
-
-void MDS::send_message_client_counted(Message *m, Session *session)
-{
-  version_t seq = session->inc_push_seq();
-  dout(10) << "send_message_client_counted " << session->info.inst.name << " seq "
-	   << seq << " " << *m << dendl;
-  if (session->connection) {
-    session->connection->send_message(m);
-  } else {
-    session->preopen_out_queue.push_back(m);
-  }
-}
-
-void MDS::send_message_client(Message *m, Session *session)
-{
-  dout(10) << "send_message_client " << session->info.inst << " " << *m << dendl;
-  if (session->connection) {
-    session->connection->send_message(m);
-  } else {
-    session->preopen_out_queue.push_back(m);
-  }
-}
-
 int MDS::init(MDSMap::DaemonState wanted_state)
 {
   dout(10) << sizeof(MDSCacheObject) << "\tMDSCacheObject" << dendl;
@@ -1315,8 +1172,6 @@ void MDS::tick()
   mds_load_t load = balancer->get_load(now);
   
   if (logger) {
-    req_rate = logger->get(l_mds_request);
-    
     logger->set(l_mds_load_cent, 100 * load.mds_load());
     logger->set(l_mds_dispatch_queue_len, messenger->get_dispatch_queue_len());
     logger->set(l_mds_subtrees, mdcache->num_subtrees());
@@ -1459,32 +1314,32 @@ int MDS::_handle_command(
   assert(outbl != NULL);
   assert(outs != NULL);
 
-  class SuicideLater : public MDSInternalContext
+  class SuicideLater : public MDSDaemonInternalContext
   {
     public:
 
-    SuicideLater(MDS *mds) : MDSInternalContext(mds) {}
+    SuicideLater(MDS *mds) : MDSDaemonInternalContext(mds) {}
     void finish(int r) {
       // Wait a little to improve chances of caller getting
       // our response before seeing us disappear from mdsmap
       sleep(1);
 
-      mds->suicide();
+      mds_daemon->suicide();
     }
   };
 
 
-  class RespawnLater : public MDSInternalContext
+  class RespawnLater : public MDSDaemonInternalContext
   {
     public:
 
-    RespawnLater(MDS *mds) : MDSInternalContext(mds) {}
+    RespawnLater(MDS *mds) : MDSDaemonInternalContext(mds) {}
     void finish(int r) {
       // Wait a little to improve chances of caller getting
       // our response before seeing us disappear from mdsmap
       sleep(1);
 
-      mds->respawn();
+      mds_daemon->respawn();
     }
   };
 
@@ -2074,11 +1929,10 @@ void MDS::request_state(MDSMap::DaemonState s)
   beacon.send();
 }
 
-
-class C_MDS_CreateFinish : public MDSInternalContext {
+class C_MDS_CreateFinish : public MDSDaemonInternalContext {
 public:
-  C_MDS_CreateFinish(MDS *m) : MDSInternalContext(m) {}
-  void finish(int r) { mds->creating_done(); }
+  C_MDS_CreateFinish(MDS *m) : MDSDaemonInternalContext(m) {}
+  void finish(int r) { mds_daemon->creating_done(); }
 };
 
 void MDS::boot_create()
@@ -2138,13 +1992,12 @@ void MDS::creating_done()
   request_state(MDSMap::STATE_ACTIVE);
 }
 
-
-class C_MDS_BootStart : public MDSInternalContext {
+class C_MDS_BootStart : public MDSDaemonInternalContext {
   MDS::BootStep nextstep;
 public:
-  C_MDS_BootStart(MDS *m, MDS::BootStep n) : MDSInternalContext(m), nextstep(n) {}
+  C_MDS_BootStart(MDS *m, MDS::BootStep n) : MDSDaemonInternalContext(m), nextstep(n) {}
   void finish(int r) {
-    mds->boot_start(nextstep, r);
+    mds_daemon->boot_start(nextstep, r);
   }
 };
 
@@ -2288,13 +2141,13 @@ void MDS::replay_start()
 }
 
 
-class MDS::C_MDS_StandbyReplayRestartFinish : public MDSIOContext {
+class MDS::C_MDS_StandbyReplayRestartFinish : public MDSDaemonIOContext {
   uint64_t old_read_pos;
 public:
   C_MDS_StandbyReplayRestartFinish(MDS *mds_, uint64_t old_read_pos_) :
-    MDSIOContext(mds_), old_read_pos(old_read_pos_) {}
+    MDSDaemonIOContext(mds_), old_read_pos(old_read_pos_) {}
   void finish(int r) {
-    mds->_standby_replay_restart_finish(r, old_read_pos);
+    mds_daemon->_standby_replay_restart_finish(r, old_read_pos);
   }
 };
 
@@ -2342,12 +2195,12 @@ inline void MDS::standby_replay_restart()
   }
 }
 
-class MDS::C_MDS_StandbyReplayRestart : public MDSInternalContext {
+class MDS::C_MDS_StandbyReplayRestart : public MDSDaemonInternalContext {
 public:
-  C_MDS_StandbyReplayRestart(MDS *m) : MDSInternalContext(m) {}
+  C_MDS_StandbyReplayRestart(MDS *m) : MDSDaemonInternalContext(m) {}
   void finish(int r) {
     assert(!r);
-    mds->standby_replay_restart();
+    mds_daemon->standby_replay_restart();
   }
 };
 
@@ -2749,25 +2602,7 @@ void MDS::respawn()
   assert(0);
 }
 
-void MDS::handle_write_error(int err)
-{
-  if (err == -EBLACKLISTED) {
-    derr << "we have been blacklisted (fenced), respawning..." << dendl;
-    respawn();
-    return;
-  }
 
-  if (g_conf->mds_action_on_write_error >= 2) {
-    derr << "unhandled write error " << cpp_strerror(err) << ", suicide..." << dendl;
-    suicide();
-  } else if (g_conf->mds_action_on_write_error == 1) {
-    derr << "unhandled write error " << cpp_strerror(err) << ", force readonly..." << dendl;
-    mdcache->force_readonly();
-  } else {
-    // ignore;
-    derr << "unhandled write error " << cpp_strerror(err) << ", ignore..." << dendl;
-  }
-}
 
 bool MDS::ms_dispatch(Message *m)
 {
@@ -2809,17 +2644,6 @@ bool MDS::ms_get_authorizer(int dest_type, AuthAuthorizer **authorizer, bool for
   *authorizer = monc->auth->build_authorizer(dest_type);
   return *authorizer != NULL;
 }
-
-
-#define ALLOW_MESSAGES_FROM(peers) \
-do { \
-  if (m->get_connection() && (m->get_connection()->get_peer_type() & (peers)) == 0) { \
-    dout(0) << __FILE__ << "." << __LINE__ << ": filtered out request, peer=" << m->get_connection()->get_peer_type() \
-           << " allowing=" << #peers << " message=" << *m << dendl; \
-    m->put();							    \
-    return true; \
-  } \
-} while (0)
 
 
 /*
@@ -2869,148 +2693,6 @@ bool MDS::handle_core_message(Message *m)
     return false;
   }
   return true;
-}
-
-/*
- * lower priority messages we defer if we seem laggy
- */
-bool MDS::handle_deferrable_message(Message *m)
-{
-  int port = m->get_type() & 0xff00;
-
-  switch (port) {
-  case MDS_PORT_CACHE:
-    ALLOW_MESSAGES_FROM(CEPH_ENTITY_TYPE_MDS);
-    mdcache->dispatch(m);
-    break;
-    
-  case MDS_PORT_MIGRATOR:
-    ALLOW_MESSAGES_FROM(CEPH_ENTITY_TYPE_MDS);
-    mdcache->migrator->dispatch(m);
-    break;
-    
-  default:
-    switch (m->get_type()) {
-      // SERVER
-    case CEPH_MSG_CLIENT_SESSION:
-    case CEPH_MSG_CLIENT_RECONNECT:
-      ALLOW_MESSAGES_FROM(CEPH_ENTITY_TYPE_CLIENT);
-      // fall-thru
-    case CEPH_MSG_CLIENT_REQUEST:
-      server->dispatch(m);
-      break;
-    case MSG_MDS_SLAVE_REQUEST:
-      ALLOW_MESSAGES_FROM(CEPH_ENTITY_TYPE_MDS);
-      server->dispatch(m);
-      break;
-      
-    case MSG_MDS_HEARTBEAT:
-      ALLOW_MESSAGES_FROM(CEPH_ENTITY_TYPE_MDS);
-      balancer->proc_message(m);
-      break;
-	  
-    case MSG_MDS_TABLE_REQUEST:
-      ALLOW_MESSAGES_FROM(CEPH_ENTITY_TYPE_MDS);
-      {
-	MMDSTableRequest *req = static_cast<MMDSTableRequest*>(m);
-	if (req->op < 0) {
-	  MDSTableClient *client = get_table_client(req->table);
-	      client->handle_request(req);
-	} else {
-	  MDSTableServer *server = get_table_server(req->table);
-	  server->handle_request(req);
-	}
-      }
-      break;
-
-    case MSG_MDS_LOCK:
-    case MSG_MDS_INODEFILECAPS:
-      ALLOW_MESSAGES_FROM(CEPH_ENTITY_TYPE_MDS);
-      locker->dispatch(m);
-      break;
-      
-    case CEPH_MSG_CLIENT_CAPS:
-    case CEPH_MSG_CLIENT_CAPRELEASE:
-    case CEPH_MSG_CLIENT_LEASE:
-      ALLOW_MESSAGES_FROM(CEPH_ENTITY_TYPE_CLIENT);
-      locker->dispatch(m);
-      break;
-      
-    default:
-      return false;
-    }
-  }
-
-  return true;
-}
-
-bool MDS::is_stale_message(Message *m)
-{
-  // from bad mds?
-  if (m->get_source().is_mds()) {
-    mds_rank_t from = mds_rank_t(m->get_source().num());
-    if (!mdsmap->have_inst(from) ||
-	mdsmap->get_inst(from) != m->get_source_inst() ||
-	mdsmap->is_down(from)) {
-      // bogus mds?
-      if (m->get_type() == CEPH_MSG_MDS_MAP) {
-	dout(5) << "got " << *m << " from old/bad/imposter mds " << m->get_source()
-		<< ", but it's an mdsmap, looking at it" << dendl;
-      } else if (m->get_type() == MSG_MDS_CACHEEXPIRE &&
-		 mdsmap->get_inst(from) == m->get_source_inst()) {
-	dout(5) << "got " << *m << " from down mds " << m->get_source()
-		<< ", but it's a cache_expire, looking at it" << dendl;
-      } else {
-	dout(5) << "got " << *m << " from down/old/bad/imposter mds " << m->get_source()
-		<< ", dropping" << dendl;
-	return true;
-      }
-    }
-  }
-  return false;
-}
-
-/**
- * Advance finished_queue and waiting_for_nolaggy.
- *
- * Usually drain both queues, but may not drain waiting_for_nolaggy
- * if beacon is currently laggy.
- */
-void MDS::_advance_queues()
-{
-  assert(mds_lock.is_locked_by_me());
-
-  while (!finished_queue.empty()) {
-    dout(7) << "mds has " << finished_queue.size() << " queued contexts" << dendl;
-    dout(10) << finished_queue << dendl;
-    list<MDSInternalContextBase*> ls;
-    ls.swap(finished_queue);
-    while (!ls.empty()) {
-      dout(10) << " finish " << ls.front() << dendl;
-      ls.front()->complete(0);
-      ls.pop_front();
-
-      heartbeat_reset();
-    }
-  }
-
-  while (!waiting_for_nolaggy.empty()) {
-    // stop if we're laggy now!
-    if (beacon.is_laggy())
-      break;
-
-    Message *old = waiting_for_nolaggy.front();
-    waiting_for_nolaggy.pop_front();
-
-    if (is_stale_message(old)) {
-      old->put();
-    } else {
-      dout(7) << " processing laggy deferred " << *old << dendl;
-      handle_deferrable_message(old);
-    }
-
-    heartbeat_reset();
-  }
 }
 
 /* If this function returns true, it has put the message. If it returns false,
@@ -3356,74 +3038,6 @@ void MDS::set_want_state(MDSMap::DaemonState newstate)
   }
 }
 
-/**
- * Call this when you take mds_lock, or periodically if you're going to
- * hold the lock for a long time (e.g. iterating over clients/inodes)
- */
-void MDS::heartbeat_reset()
-{
-  // Any thread might jump into mds_lock and call us immediately
-  // after a call to suicide() completes, in which case MDS::hb
-  // has been freed and we are a no-op.
-  if (!hb) {
-      assert(want_state == CEPH_MDS_STATE_DNE);
-      return;
-  }
-
-  // NB not enabling suicide grace, because the mon takes care of killing us
-  // (by blacklisting us) when we fail to send beacons, and it's simpler to
-  // only have one way of dying.
-  cct->get_heartbeat_map()->reset_timeout(hb, g_conf->mds_beacon_grace, 0);
-}
 
 
-void *MDS::ProgressThread::entry()
-{
-  Mutex::Locker l(mds->mds_lock);
-  while (true) {
-    while (!mds->stopping &&
-	   mds->finished_queue.empty() &&
-	   (mds->waiting_for_nolaggy.empty() || mds->beacon.is_laggy())) {
-      cond.Wait(mds->mds_lock);
-    }
 
-    if (mds->stopping) {
-      break;
-    }
-
-    mds->_advance_queues();
-  }
-
-  return NULL;
-}
-
-
-void MDS::ProgressThread::shutdown()
-{
-  assert(mds->mds_lock.is_locked_by_me());
-  assert(mds->stopping);
-
-  if (am_self()) {
-    // Stopping is set, we will fall out of our main loop naturally
-  } else {
-    // Kick the thread to notice mds->stopping, and join it
-    cond.Signal();
-    mds->mds_lock.Unlock();
-    if (is_started())
-      join();
-    mds->mds_lock.Lock();
-  }
-}
-
-/**
- * This is used whenever a RADOS operation has been cancelled
- * or a RADOS client has been blacklisted, to cause the MDS and
- * any clients to wait for this OSD epoch before using any new caps.
- *
- * See doc/cephfs/eviction
- */
-void MDS::set_osd_epoch_barrier(epoch_t e)
-{
-  dout(4) << __func__ << ": epoch=" << e << dendl;
-  osd_epoch_barrier = e;
-}
