+++ src/mds/MDS.h
@@ -33,76 +33,14 @@
 #include "common/Finisher.h"
 #include "common/cmdparse.h"
 
+#include "MDSRank.h"
 #include "MDSMap.h"
 
-#include "SessionMap.h"
 #include "Beacon.h"
 
 
 #define CEPH_MDS_PROTOCOL    27 /* cluster internal */
 
-enum {
-  l_mds_first = 2000,
-  l_mds_request,
-  l_mds_reply,
-  l_mds_reply_latency,
-  l_mds_forward,
-  l_mds_dir_fetch,
-  l_mds_dir_commit,
-  l_mds_dir_split,
-  l_mds_inode_max,
-  l_mds_inodes,
-  l_mds_inodes_top,
-  l_mds_inodes_bottom,
-  l_mds_inodes_pin_tail,
-  l_mds_inodes_pinned,
-  l_mds_inodes_expired,
-  l_mds_inodes_with_caps,
-  l_mds_caps,
-  l_mds_subtrees,
-  l_mds_traverse,
-  l_mds_traverse_hit,
-  l_mds_traverse_forward,
-  l_mds_traverse_discover,
-  l_mds_traverse_dir_fetch,
-  l_mds_traverse_remote_ino,
-  l_mds_traverse_lock,
-  l_mds_load_cent,
-  l_mds_dispatch_queue_len,
-  l_mds_exported,
-  l_mds_exported_inodes,
-  l_mds_imported,
-  l_mds_imported_inodes,
-  l_mds_last,
-};
-
-// memory utilization
-enum {
-  l_mdm_first = 2500,
-  l_mdm_ino,
-  l_mdm_inoa,
-  l_mdm_inos,
-  l_mdm_dir,
-  l_mdm_dira,
-  l_mdm_dirs,
-  l_mdm_dn,
-  l_mdm_dna,
-  l_mdm_dns,
-  l_mdm_cap,
-  l_mdm_capa,
-  l_mdm_caps,
-  l_mdm_rss,
-  l_mdm_heap,
-  l_mdm_malloc,
-  l_mdm_buf,
-  l_mdm_last,
-};
-
-
-
-namespace ceph {
-  struct heartbeat_handle_d;
-}
 class filepath;
 
 class MonClient;
@@ -113,7 +51,6 @@ class Filer;
 class Server;
 class Locker;
 class MDCache;
-class MDLog;
 class MDBalancer;
 class MDSInternalContextBase;
 
@@ -123,9 +60,6 @@ class CDentry;
 
 class Messenger;
 class Message;
-
-class MClientRequest;
-class MClientReply;
 
 class MMDSBeacon;
 
@@ -138,7 +72,7 @@ class MDSTableClient;
 
 class AuthAuthorizeHandlerRegistry;
 
-class MDS : public Dispatcher, public md_config_obs_t {
+class MDS : public MDSRank, public Dispatcher, public md_config_obs_t {
  public:
 
   /* Global MDS lock: every time someone takes this, they must
@@ -150,20 +84,18 @@ class MDS : public Dispatcher, public md_config_obs_t {
 
   SafeTimer    timer;
 
- private:
-  ceph::heartbeat_handle_d *hb;  // Heartbeat for threads using mds_lock
-  void heartbeat_reset();
+ public:
+  // XXX fixme beacon only public for the benefit of MDSRank who wants
+  // to see if it's laggy
   Beacon  beacon;
+ private:
   void set_want_state(MDSMap::DaemonState newstate);
  public:
-  utime_t get_laggy_until() {return beacon.get_laggy_until();}
 
   AuthAuthorizeHandlerRegistry *authorize_handler_cluster_registry;
   AuthAuthorizeHandlerRegistry *authorize_handler_service_registry;
 
   string name;
-  mds_rank_t whoami;
-  int incarnation;
 
   mds_rank_t standby_for_rank;
   MDSMap::DaemonState standby_type;  // one of STANDBY_REPLAY, ONESHOT_REPLAY
@@ -177,147 +109,31 @@ class MDS : public Dispatcher, public md_config_obs_t {
   LogClient    log_client;
   LogChannelRef clog;
 
-  // sub systems
-  Server       *server;
-  MDCache      *mdcache;
-  Locker       *locker;
-  MDLog        *mdlog;
-  MDBalancer   *balancer;
-
-  InoTable     *inotable;
-
-  SnapServer   *snapserver;
-  SnapClient   *snapclient;
-
-  MDSTableClient *get_table_client(int t);
-  MDSTableServer *get_table_server(int t);
-
-  PerfCounters       *logger, *mlogger;
-  OpTracker    op_tracker;
+  //OpTracker    op_tracker;
 
   Finisher finisher;
 
   int orig_argc;
   const char **orig_argv;
 
- protected:
-  // -- MDS state --
-  MDSMap::DaemonState last_state;
-  MDSMap::DaemonState state;         // my confirmed state
-  MDSMap::DaemonState want_state;    // the state i want
-
-  list<MDSInternalContextBase*> waiting_for_active, waiting_for_replay, waiting_for_reconnect, waiting_for_resolve;
-  list<MDSInternalContextBase*> replay_queue;
-  map<mds_rank_t, list<MDSInternalContextBase*> > waiting_for_active_peer;
-  list<Message*> waiting_for_nolaggy;
-  map<epoch_t, list<MDSInternalContextBase*> > waiting_for_mdsmap;
-
-  map<mds_rank_t, version_t> peer_mdsmap_epoch;
-
-  ceph_tid_t last_tid;    // for mds-initiated requests (e.g. stray rename)
-
-  epoch_t osd_epoch_barrier;
-
- public:
-  void set_osd_epoch_barrier(epoch_t e);
-  epoch_t get_osd_epoch_barrier() const {return osd_epoch_barrier;}
-
-  void wait_for_active(MDSInternalContextBase *c) { 
-    waiting_for_active.push_back(c); 
-  }
-  void wait_for_active_peer(mds_rank_t who, MDSInternalContextBase *c) { 
-    waiting_for_active_peer[who].push_back(c);
-  }
-  void wait_for_replay(MDSInternalContextBase *c) { 
-    waiting_for_replay.push_back(c); 
-  }
-  void wait_for_reconnect(MDSInternalContextBase *c) {
-    waiting_for_reconnect.push_back(c);
-  }
-  void wait_for_resolve(MDSInternalContextBase *c) {
-    waiting_for_resolve.push_back(c);
-  }
-  void wait_for_mdsmap(epoch_t e, MDSInternalContextBase *c) {
-    waiting_for_mdsmap[e].push_back(c);
-  }
-  void enqueue_replay(MDSInternalContextBase *c) {
-    replay_queue.push_back(c);
-  }
-
-  MDSMap::DaemonState get_state() { return state; } 
-  MDSMap::DaemonState get_want_state() { return want_state; } 
-  bool is_creating() { return state == MDSMap::STATE_CREATING; }
-  bool is_starting() { return state == MDSMap::STATE_STARTING; }
-  bool is_standby()  { return state == MDSMap::STATE_STANDBY; }
-  bool is_replay()   { return state == MDSMap::STATE_REPLAY; }
-  bool is_standby_replay() { return state == MDSMap::STATE_STANDBY_REPLAY; }
-  bool is_resolve()  { return state == MDSMap::STATE_RESOLVE; }
-  bool is_reconnect() { return state == MDSMap::STATE_RECONNECT; }
-  bool is_rejoin()   { return state == MDSMap::STATE_REJOIN; }
-  bool is_clientreplay()   { return state == MDSMap::STATE_CLIENTREPLAY; }
-  bool is_active()   { return state == MDSMap::STATE_ACTIVE; }
-  bool is_stopping() { return state == MDSMap::STATE_STOPPING; }
-
-  bool is_oneshot_replay()   { return state == MDSMap::STATE_ONESHOT_REPLAY; }
-  bool is_any_replay() { return (is_replay() || is_standby_replay() ||
-                                 is_oneshot_replay()); }
-
-  bool is_stopped()  { return mdsmap->is_stopped(whoami); }
-
+public:
   void request_state(MDSMap::DaemonState s);
 
-  ceph_tid_t issue_tid() { return ++last_tid; }
-    
-
-  // -- waiters --
-private:
-  list<MDSInternalContextBase*> finished_queue;
-  void _advance_queues();
-public:
-
-  void queue_waiter(MDSInternalContextBase *c) {
-    finished_queue.push_back(c);
-    progress_thread.signal();
-  }
-  void queue_waiters(list<MDSInternalContextBase*>& ls) {
-    finished_queue.splice( finished_queue.end(), ls );
-    progress_thread.signal();
-  }
-  bool queue_one_replay() {
-    if (replay_queue.empty())
-      return false;
-    queue_waiter(replay_queue.front());
-    replay_queue.pop_front();
-    return true;
-  }
-  
   // tick and other timer fun
   class C_MDS_Tick : public MDSInternalContext {
+    protected:
+      MDS *mds_daemon;
   public:
-    C_MDS_Tick(MDS *m) : MDSInternalContext(m) {}
+    C_MDS_Tick(MDS *m) : MDSInternalContext(m), mds_daemon(m) {}
     void finish(int r) {
-      mds->tick_event = 0;
-      mds->tick();
+      mds_daemon->tick_event = 0;
+      mds_daemon->tick();
     }
   } *tick_event;
   void     reset_tick();
 
   // -- client map --
-  SessionMap   sessionmap;
   epoch_t      last_client_mdsmap_bcast;
-  //void log_clientmap(Context *c);
-
-
-  // shutdown crap
-  int req_rate;
-
-  // ino's and fh's
- public:
-
-  int get_req_rate() { return req_rate; }
-  Session *get_session(client_t client) {
-    return sessionmap.get_session(entity_name_t::CLIENT(client.v));
-  }
 
  private:
   int dispatch_depth;
@@ -331,44 +147,12 @@ public:
   bool ms_handle_reset(Connection *con);
   void ms_handle_remote_reset(Connection *con);
 
-private:
-  class ProgressThread : public Thread {
-    MDS *mds;
-    Cond cond;
-  public:
-    ProgressThread(MDS *mds_) : mds(mds_) {}
-    void * entry(); 
-    void shutdown();
-    void signal() {cond.Signal();}
-  } progress_thread;
-  void _progress_thread();
-
  public:
   MDS(const std::string &n, Messenger *m, MonClient *mc);
   ~MDS();
 
   // handle a signal (e.g., SIGTERM)
   void handle_signal(int signum);
-
-  // who am i etc
-  mds_rank_t get_nodeid() const { return whoami; }
-  uint64_t get_metadata_pool() { return mdsmap->get_metadata_pool(); }
-  MDSMap *get_mds_map() { return mdsmap; }
-
-  void send_message_mds(Message *m, mds_rank_t mds);
-  void forward_message_mds(Message *req, mds_rank_t mds);
-
-  void send_message_client_counted(Message *m, client_t client);
-  void send_message_client_counted(Message *m, Session *session);
-  void send_message_client_counted(Message *m, Connection *connection);
-  void send_message_client_counted(Message *m, const ConnectionRef& con) {
-    send_message_client_counted(m, con.get());
-  }
-  void send_message_client(Message *m, Session *session);
-  void send_message(Message *m, Connection *c);
-  void send_message(Message *m, const ConnectionRef& c) {
-    send_message(m, c.get());
-  }
 
   // start up, shutdown
   int init(MDSMap::DaemonState wanted_state=MDSMap::STATE_BOOT);
@@ -502,11 +286,9 @@ private:
    * this process was run with, then terminate this process
    */
   void respawn();
-  void handle_write_error(int err);
 
   void tick();
   
-
   void inc_dispatch_depth() { ++dispatch_depth; }
   void dec_dispatch_depth() { --dispatch_depth; }
 
@@ -514,10 +296,7 @@ private:
   bool _dispatch(Message *m, bool new_msg);
 
   protected:
-  bool is_stale_message(Message *m);
-
   bool handle_core_message(Message *m);
-  bool handle_deferrable_message(Message *m);
   
   // special message types
   int _handle_command_legacy(std::vector<std::string> args);
@@ -532,22 +311,5 @@ private:
   void handle_mds_map(class MMDSMap *m);
 };
 
-
-/* This expects to be given a reference which it is responsible for.
- * The finish function calls functions which
- * will put the Message exactly once.*/
-class C_MDS_RetryMessage : public MDSInternalContext {
-  Message *m;
-public:
-  C_MDS_RetryMessage(MDS *mds, Message *m) : MDSInternalContext(mds) {
-    assert(m);
-    this->m = m;
-  }
-  virtual void finish(int r) {
-    mds->inc_dispatch_depth();
-    mds->_dispatch(m, false);
-    mds->dec_dispatch_depth();
-  }
-};
 
 #endif
